[apps/reg] Model::numberOfCoefficients return an int, not a double

+++ apps/regression/model/exponential_model.h
@@ -14,7 +14,7 @@ public:
   double evaluate(double * modelCoefficients, double x) const override;
   double levelSet(double * modelCoefficients, double y) const override;
   double partialDerivate(double * modelCoefficients, int derivateCoefficientIndex, double x) const override;
-  double numberOfCoefficients() const override { return 2; }
+  int numberOfCoefficients() const override { return 2; }
 };
 
 }
